+++ imod/mf6/__init__.py
@@ -12,6 +12,7 @@ from imod.mf6.drn import Drainage
 from imod.mf6.dsp import Dispersion
 from imod.mf6.evt import Evapotranspiration
 from imod.mf6.ghb import GeneralHeadBoundary
+from imod.mf6.hfb import HorizontalFlowBarrier
 from imod.mf6.ic import InitialConditions
 from imod.mf6.ims import (
     Solution,