+++ imod/tests/test_wq/test_wq_wel.py
@@ -10,29 +10,40 @@ from imod.wq import Well
 
 @pytest.fixture(scope="module")
 def well():
+    """Five separate wells, each starting at a different time."""
     datetimes = pd.date_range("2000-01-01", "2000-01-05")
     y = np.arange(4.5, 0.0, -1.0)
     x = np.arange(0.5, 5.0, 1.0)
-    return Well(id_name="well", x=x, y=y, rate=5.0, layer=2, time=datetimes)
+    id_name = [f"well_{i}" for i in range(x.size)]
+    return Well(id_name=id_name, x=x, y=y, rate=5.0, layer=2, time=datetimes)
 
 
 @pytest.fixture(scope="module")
 def well_conc():
+    """
+    Five separate wells, each starting at a different time, with a single
+    species.
+    """
     datetimes = pd.date_range("2000-01-01", "2000-01-05")
     y = np.arange(4.5, 0.0, -1.0)
     x = np.arange(0.5, 5.0, 1.0)
+    id_name = [f"well_{i}" for i in range(x.size)]
     return Well(
-        id_name="well", x=x, y=y, rate=5.0, layer=2, time=datetimes, concentration=2.5
+        id_name=id_name, x=x, y=y, rate=5.0, layer=2, time=datetimes, concentration=2.5
     )
 
 
 @pytest.fixture(scope="module")
 def well_conc_multiple_species():
+    """
+    Five separate wells, each starting at a different time, with two species.
+    """
     datetimes = pd.date_range("2000-01-01", "2000-01-05")
     y = np.arange(4.5, 0.0, -1.0)
     x = np.arange(0.5, 5.0, 1.0)
+    id_name = [f"well_{i}" for i in range(x.size)]
     w = Well(
-        id_name="well", x=x, y=y, rate=5.0, layer=2, time=datetimes, concentration=2.5
+        id_name=id_name, x=x, y=y, rate=5.0, layer=2, time=datetimes, concentration=2.5
     )
     conc1 = w["concentration"].assign_coords(species=1)
     conc2 = w["concentration"].assign_coords(species=2)
@@ -44,11 +55,7 @@ def test_render(well):
     wel = well
     directory = pathlib.Path("well")
     compare = """
-    wel_p1_s1_l2 = well/well_20000101000000_l2.ipf
-    wel_p2_s1_l2 = well/well_20000102000000_l2.ipf
-    wel_p3_s1_l2 = well/well_20000103000000_l2.ipf
-    wel_p4_s1_l2 = well/well_20000104000000_l2.ipf
-    wel_p5_s1_l2 = well/well_20000105000000_l2.ipf"""
+    wel_p?_s1_l2 = well/rate_l2.ipf"""
 
     actual = wel._render(
         directory, globaltimes=wel["time"].values, system_index=1, nlayer=3
@@ -66,7 +73,7 @@ def test_render__notime_nolayer(well):
     wel = Well(**d)
     directory = pathlib.Path("well")
     compare = """
-    wel_p?_s1_l? = well/well.ipf"""
+    wel_p?_s1_l$ = well/rate.ipf"""
 
     actual = wel._render(directory, globaltimes=["?"], system_index=1, nlayer=3)
     assert actual == compare
@@ -78,7 +85,7 @@ def test_render__notime_layer(well):
     wel = Well(**d)
     directory = pathlib.Path("well")
     compare = """
-    wel_p?_s1_l1:5 = well/well_l:.ipf"""
+    wel_p?_s1_l1:5 = well/rate_l:.ipf"""
 
     actual = wel._render(directory, globaltimes=["?"], system_index=1, nlayer=3)
     assert actual == compare
@@ -89,11 +96,7 @@ def test_render__time_nolayer(well):
     wel = Well(**d)
     directory = pathlib.Path("well")
     compare = """
-    wel_p1_s1_l? = well/well_20000101000000.ipf
-    wel_p2_s1_l? = well/well_20000102000000.ipf
-    wel_p3_s1_l? = well/well_20000103000000.ipf
-    wel_p4_s1_l? = well/well_20000104000000.ipf
-    wel_p5_s1_l? = well/well_20000105000000.ipf"""
+    wel_p?_s1_l$ = well/rate.ipf"""
 
     assert (
         wel._render(directory, globaltimes=wel["time"].values, system_index=1, nlayer=3)
@@ -107,64 +110,11 @@ def test_render__time_layer(well):
     wel = Well(**d)
     directory = pathlib.Path("well")
     compare = """
-    wel_p1_s1_l1 = well/well_20000101000000_l1.ipf
-    wel_p2_s1_l2 = well/well_20000102000000_l2.ipf
-    wel_p3_s1_l3 = well/well_20000103000000_l3.ipf
-    wel_p4_s1_l4 = well/well_20000104000000_l4.ipf
-    wel_p5_s1_l5 = well/well_20000105000000_l5.ipf"""
+    wel_p?_s1_l1:5 = well/rate_l:.ipf"""
 
     actual = wel._render(
         directory, globaltimes=wel["time"].values, system_index=1, nlayer=3
     )
-    assert actual == compare
-
-
-def test_stress_repeats__single_layer():
-    datetimes = pd.date_range("2000-01-01", "2000-01-05")
-    x = [1.0, 1.0]
-    y = [1.0, 1.0]
-    time = [datetimes[0], datetimes[1]]
-    wel = Well(id_name="well", x=x, y=y, rate=5.0, layer=1, time=time)
-    # Set a periodic boundary condition
-    stress_repeats = {
-        datetimes[2]: datetimes[0],
-        datetimes[3]: datetimes[1],
-        datetimes[4]: datetimes[0],
-    }
-    wel.repeat_stress(stress_repeats)
-
-    directory = pathlib.Path("well")
-    compare = """
-    wel_p1_s1_l1 = well/well_20000101000000_l1.ipf
-    wel_p2_s1_l1 = well/well_20000102000000_l1.ipf
-    wel_p3_s1_l1 = well/well_20000101000000_l1.ipf
-    wel_p4_s1_l1 = well/well_20000102000000_l1.ipf
-    wel_p5_s1_l1 = well/well_20000101000000_l1.ipf"""
-
-    actual = wel._render(directory, globaltimes=datetimes, system_index=1, nlayer=3)
-    assert actual == compare
-
-
-def test_stress_repeats__multiple_layers():
-    datetimes = pd.date_range("2000-01-01", "2000-01-04")
-    layer = [1, 1, 2]
-    x = [1.0, 1.0, 1.0]
-    y = [1.0, 1.0, 1.0]
-    time = [datetimes[0], datetimes[1], datetimes[0]]
-    wel = Well(id_name="well", x=x, y=y, rate=5.0, layer=layer, time=time)
-    # Set a periodic boundary condition
-    # test number of well layers is not constant between periods
-    stress_repeats = {datetimes[2]: datetimes[0], datetimes[3]: datetimes[1]}
-    wel.repeat_stress(stress_repeats)
-
-    directory = pathlib.Path("well")
-    compare = """
-    wel_p1_s1_l1:2 = well/well_20000101000000_l:.ipf
-    wel_p2_s1_l1 = well/well_20000102000000_l1.ipf
-    wel_p3_s1_l1:2 = well/well_20000101000000_l:.ipf
-    wel_p4_s1_l1 = well/well_20000102000000_l1.ipf"""
-
-    actual = wel._render(directory, globaltimes=datetimes, system_index=1, nlayer=3)
     assert actual == compare
 
 
@@ -173,11 +123,7 @@ def test_render_concentration(well_conc, tmp_path):
     directory = pathlib.Path("well")
     # Test rate
     compare = """
-    wel_p1_s1_l2 = well/well_20000101000000_l2.ipf
-    wel_p2_s1_l2 = well/well_20000102000000_l2.ipf
-    wel_p3_s1_l2 = well/well_20000103000000_l2.ipf
-    wel_p4_s1_l2 = well/well_20000104000000_l2.ipf
-    wel_p5_s1_l2 = well/well_20000105000000_l2.ipf"""
+    wel_p?_s1_l2 = well/rate_l2.ipf"""
 
     actual = wel._render(
         directory, globaltimes=wel["time"].values, system_index=1, nlayer=3
@@ -186,11 +132,7 @@ def test_render_concentration(well_conc, tmp_path):
 
     # Test concentration
     compare = """
-    cwel_t1_p1_l2 = well/well-concentration_20000101000000_l2.ipf
-    cwel_t1_p2_l2 = well/well-concentration_20000102000000_l2.ipf
-    cwel_t1_p3_l2 = well/well-concentration_20000103000000_l2.ipf
-    cwel_t1_p4_l2 = well/well-concentration_20000104000000_l2.ipf
-    cwel_t1_p5_l2 = well/well-concentration_20000105000000_l2.ipf"""
+    cwel_t1_p?_l2 = well/concentration_l2.ipf"""
 
     actual = wel._render_ssm(directory, globaltimes=wel["time"].values, nlayer=3)
     assert actual == compare
@@ -201,11 +143,7 @@ def test_render_concentration_multiple_species(well_conc_multiple_species, tmp_p
     directory = pathlib.Path("well")
     # Test rate, not affected by multiple species
     compare = """
-    wel_p1_s1_l2 = well/well_20000101000000_l2.ipf
-    wel_p2_s1_l2 = well/well_20000102000000_l2.ipf
-    wel_p3_s1_l2 = well/well_20000103000000_l2.ipf
-    wel_p4_s1_l2 = well/well_20000104000000_l2.ipf
-    wel_p5_s1_l2 = well/well_20000105000000_l2.ipf"""
+    wel_p?_s1_l2 = well/rate_l2.ipf"""
 
     actual = wel._render(
         directory, globaltimes=wel["time"].values, system_index=1, nlayer=3
@@ -214,16 +152,8 @@ def test_render_concentration_multiple_species(well_conc_multiple_species, tmp_p
 
     # Test concentration
     compare = """
-    cwel_t1_p1_l2 = well/well-concentration_c1_20000101000000_l2.ipf
-    cwel_t1_p2_l2 = well/well-concentration_c1_20000102000000_l2.ipf
-    cwel_t1_p3_l2 = well/well-concentration_c1_20000103000000_l2.ipf
-    cwel_t1_p4_l2 = well/well-concentration_c1_20000104000000_l2.ipf
-    cwel_t1_p5_l2 = well/well-concentration_c1_20000105000000_l2.ipf
-    cwel_t2_p1_l2 = well/well-concentration_c2_20000101000000_l2.ipf
-    cwel_t2_p2_l2 = well/well-concentration_c2_20000102000000_l2.ipf
-    cwel_t2_p3_l2 = well/well-concentration_c2_20000103000000_l2.ipf
-    cwel_t2_p4_l2 = well/well-concentration_c2_20000104000000_l2.ipf
-    cwel_t2_p5_l2 = well/well-concentration_c2_20000105000000_l2.ipf"""
+    cwel_t1_p?_l2 = well/concentration_c1_l2.ipf
+    cwel_t2_p?_l2 = well/concentration_c2_l2.ipf"""
 
     actual = wel._render_ssm(directory, globaltimes=wel["time"].values, nlayer=3)
     assert actual == compare
@@ -234,11 +164,12 @@ def test_save(well, tmp_path):
     wel.save(tmp_path / "well")
 
     files = [
-        "well_20000101000000_l2.ipf",
-        "well_20000102000000_l2.ipf",
-        "well_20000103000000_l2.ipf",
-        "well_20000104000000_l2.ipf",
-        "well_20000105000000_l2.ipf",
+        "rate_l2.ipf",
+        "rate_l2/well_0.txt",
+        "rate_l2/well_1.txt",
+        "rate_l2/well_2.txt",
+        "rate_l2/well_3.txt",
+        "rate_l2/well_4.txt",
     ]
     for file in files:
         assert (tmp_path / "well" / file).is_file()
@@ -249,16 +180,18 @@ def test_save_concentration(well_conc, tmp_path):
     wel.save(tmp_path / "well")
 
     files = [
-        "well_20000101000000_l2.ipf",
-        "well_20000102000000_l2.ipf",
-        "well_20000103000000_l2.ipf",
-        "well_20000104000000_l2.ipf",
-        "well_20000105000000_l2.ipf",
-        "well-concentration_20000101000000_l2.ipf",
-        "well-concentration_20000102000000_l2.ipf",
-        "well-concentration_20000103000000_l2.ipf",
-        "well-concentration_20000104000000_l2.ipf",
-        "well-concentration_20000105000000_l2.ipf",
+        "rate_l2.ipf",
+        "concentration_l2.ipf",
+        "rate_l2/well_0.txt",
+        "rate_l2/well_1.txt",
+        "rate_l2/well_2.txt",
+        "rate_l2/well_3.txt",
+        "rate_l2/well_4.txt",
+        "concentration_l2/well_0.txt",
+        "concentration_l2/well_1.txt",
+        "concentration_l2/well_2.txt",
+        "concentration_l2/well_3.txt",
+        "concentration_l2/well_4.txt",
     ]
     for file in files:
         assert (tmp_path / "well" / file).is_file()
@@ -269,21 +202,24 @@ def test_save_multiple_species(well_conc_multiple_species, tmp_path):
     wel.save(tmp_path / "well")
 
     files = [
-        "well_20000101000000_l2.ipf",
-        "well_20000102000000_l2.ipf",
-        "well_20000103000000_l2.ipf",
-        "well_20000104000000_l2.ipf",
-        "well_20000105000000_l2.ipf",
-        "well-concentration_c1_20000101000000_l2.ipf",
-        "well-concentration_c1_20000102000000_l2.ipf",
-        "well-concentration_c1_20000103000000_l2.ipf",
-        "well-concentration_c1_20000104000000_l2.ipf",
-        "well-concentration_c1_20000105000000_l2.ipf",
-        "well-concentration_c2_20000101000000_l2.ipf",
-        "well-concentration_c2_20000102000000_l2.ipf",
-        "well-concentration_c2_20000103000000_l2.ipf",
-        "well-concentration_c2_20000104000000_l2.ipf",
-        "well-concentration_c2_20000105000000_l2.ipf",
+        "rate_l2.ipf",
+        "rate_l2/well_0.txt",
+        "rate_l2/well_1.txt",
+        "rate_l2/well_2.txt",
+        "rate_l2/well_3.txt",
+        "rate_l2/well_4.txt",
+        "concentration_c1_l2.ipf",
+        "concentration_c2_l2.ipf",
+        "concentration_c1_l2/well_0.txt",
+        "concentration_c1_l2/well_1.txt",
+        "concentration_c1_l2/well_2.txt",
+        "concentration_c1_l2/well_3.txt",
+        "concentration_c1_l2/well_4.txt",
+        "concentration_c2_l2/well_0.txt",
+        "concentration_c2_l2/well_1.txt",
+        "concentration_c2_l2/well_2.txt",
+        "concentration_c2_l2/well_3.txt",
+        "concentration_c2_l2/well_4.txt",
     ]
     for file in files:
         assert (tmp_path / "well" / file).is_file()
@@ -295,11 +231,18 @@ def test_save__time_nolayer(well, tmp_path):
     wel.save(tmp_path / "well")
 
     files = [
-        "well_20000101000000.ipf",
-        "well_20000102000000.ipf",
-        "well_20000103000000.ipf",
-        "well_20000104000000.ipf",
-        "well_20000105000000.ipf",
+        "rate.ipf",
+        "well_0.txt",
+        "well_1.txt",
+        "well_2.txt",
+        "well_3.txt",
+        "well_4.txt",
     ]
     for file in files:
         assert (tmp_path / "well" / file).is_file()
+
+
+def test_repeat_stress_error(well):
+    wel = well
+    with pytest.raises(NotImplementedError):
+        wel.repeat_stress({})